@@ -24,47 +24,57 @@
   },
   "homepage": "https://github.com/parkers12/otus_homework_project#readme",
   "devDependencies": {
+    "@babel/cli": "^7.17.10",
     "@babel/core": "^7.16.0",
-    "@babel/plugin-transform-runtime": "^7.15.8",
-    "@babel/preset-env": "^7.16.0",
-    "@babel/preset-react": "^7.16.7",
-    "@babel/preset-typescript": "^7.16.7",
+    "@babel/eslint-parser": "^7.17.0",
+    "@babel/plugin-syntax-jsx": "^7.17.12",
+    "@babel/plugin-transform-runtime": "^7.17.12",
+    "@babel/preset-env": "^7.17.12",
+    "@babel/preset-react": "^7.17.12",
+    "@svgr/webpack": "^6.2.1",
     "@types/jest": "^26.0.24",
-    "autoprefixer": "^10.3.7",
-    "babel-jest": "^27.3.1",
-    "babel-loader": "^8.2.2",
+    "autoprefixer": "^10.4.7",
+    "babel-cli": "^6.26.0",
+    "babel-core": "^6.26.3",
+    "babel-jest": "^27.4.2",
+    "babel-loader": "^8.2.3",
+    "babel-preset-react": "^6.24.1",
     "clean-webpack-plugin": "^4.0.0",
     "copy-webpack-plugin": "^9.0.1",
-    "css-loader": "^6.4.0",
+    "css-loader": "^6.5.1",
     "css-mqpacker": "^7.0.0",
     "cssnano": "^5.0.8",
-    "eslint": "^7.32.0",
-    "eslint-config-airbnb-base": "^14.2.1",
-    "eslint-config-prettier": "^8.3.0",
-    "eslint-plugin-import": "^2.24.2",
-    "eslint-plugin-jest": "^24.7.0",
+    "eslint": "^8.15.0",
+    "eslint-config-airbnb-base": "^15.0.0",
+    "eslint-config-prettier": "^8.5.0",
+    "eslint-plugin-eslint-plugin": "^4.2.0",
+    "eslint-plugin-import": "^2.26.0",
+    "eslint-plugin-jest": "^26.2.2",
+    "eslint-plugin-react": "^7.29.4",
+    "eslint-webpack-plugin": "^3.1.1",
     "file-include-webpack-plugin": "^1.3.9",
+    "file-loader": "^6.2.0",
     "html-loader": "2.1.1",
     "html-webpack-plugin": "^5.5.0",
     "husky": "^7.0.4",
-    "jest": "^27.3.1",
+    "jest": "^27.4.3",
     "lint-staged": "^11.2.6",
-    "mini-css-extract-plugin": "^2.4.2",
+    "mini-css-extract-plugin": "^2.4.5",
     "node-sass": "^7.0.1",
-    "postcss": "^8.3.9",
-    "postcss-loader": "^6.2.0",
+    "postcss": "^8.4.4",
+    "postcss-loader": "^6.2.1",
     "postcss-preset-env": "^7.5.0",
     "prettier": "^2.4.1",
     "sass": "^1.43.2",
-    "sass-loader": "^12.2.0",
-    "style-loader": "^3.3.0",
+    "sass-loader": "^12.3.0",
+    "style-loader": "^3.3.1",
     "stylelint": "^14.0.1",
     "stylelint-config-standard-scss": "^2.0.0",
     "stylelint-scss": "^4.0.0",
-    "typescript": "^4.6.4",
-    "webpack": "^5.58.2",
+    "webpack": "^5.72.0",
     "webpack-cli": "^4.9.0",
-    "webpack-dev-server": "^4.3.1"
+    "webpack-dev-server": "^4.9.0",
+    "webpack-notifier": "^1.15.0"
   },
   "husky": {
     "hooks": {
@@ -72,21 +82,27 @@
     }
   },
   "lint-staged": {
-    "*.js, .tsx": "eslint --fix",
-    "*.{js,css,scss,md,json,html}": "prettier --write"
+    "*.js, .jsx": "eslint --fix",
+    "*.{js,jsx,css,scss,md,json,html}": "prettier --write"
   },
-  "browserslist": [
-    "> 1%",
-    "last 2 versions",
-    "not dead"
-  ],
+  "browserslist": {
+    "production": [
+      ">0.2%",
+      "not dead",
+      "not op_mini all"
+    ],
+    "development": [
+      "last 1 chrome version",
+      "last 1 firefox version",
+      "last 1 safari version"
+    ]
+  },
   "dependencies": {
-    "@types/react": "^18.0.9",
-    "@types/react-dom": "^18.0.3",
     "classnames": "^2.2.6",
+    "prop-types": "^15.8.1",
     "react": "^18.1.0",
     "react-dom": "^18.1.0",
-    "react-router-dom": "^5.2.0"
+    "react-router-dom": "^6.3.0"
   },
   "stats": {
     "children": true
