@@ -1,29 +1,41 @@
 module.exports = {
-    env: {
-      browser: true,
-      es2021: true,
-      "jest/globals": true,
+  env: {
+    browser: true,
+    es6: true,
+    "jest/globals": true,
+  },
+  extends: [
+    "airbnb-base",
+    "prettier",
+    "eslint:recommended",
+    "plugin:react/recommended",
+  ],
+  plugins: ["jest", "react"],
+  parserOptions: {
+    ecmaVersion: 12,
+    sourceType: "module",
+    ecmaFeatures: {
+      jsx: true,
     },
-    extends: ["airbnb-base", "prettier"],
-    plugins: ["jest"],
-    parserOptions: {
-      ecmaVersion: 12,
-      sourceType: "module",
-      allowImportExportEverywhere: false,
-      codeFrame: true
-    },
-    parser: "babel-eslint",
-    rules: {
-      strict: 0,
-      "max-len": [
-        "error",
-        {
-          ignoreComments: true,
-        },
-      ],
-      "no-console": "off",
-      "no-alert": "off",
-      "import/prefer-default-export": "off",
-      "import/no-extraneous-dependencies": ["error", { devDependencies: true }],
-    },
-  };
+  },
+  parser: "@babel/eslint-parser",
+  rules: {
+    strict: 0,
+    "max-len": [
+      "error",
+      {
+        ignoreComments: true,
+      },
+    ],
+    "react/no-set-state": "off",
+    "no-console": "off",
+    "no-alert": "off",
+    "import/prefer-default-export": "off",
+    "import/no-extraneous-dependencies": [
+      "error",
+      {
+        devDependencies: true,
+      },
+    ],
+  },
+};
