@@ -3,18 +3,17 @@ module.exports = (api) => ({
     [
       "@babel/preset-env",
       {
-        targets: api.caller((caller) => caller && caller.target === "node") ?
-          {
-            node: "current"
-          } :
-          {
-            chrome: "58",
-            ie: "11"
-          },
-      }
+        targets: api.caller((caller) => caller && caller.target === "node")
+          ? {
+              node: "current",
+            }
+          : {
+              chrome: "58",
+              ie: "11",
+            },
+      },
     ],
-    '@babel/preset-react',
-    "@babel/preset-typescript"
+    "@babel/preset-react",
   ],
 
   plugins: ["@babel/plugin-transform-runtime"],
