@@ -1,9 +1,9 @@
 const dbConfig = {
-    HOST: "localhost",
-    USER: "root",
-    PASSWORD: "luckshimi2!W",
-    PORT: 3306,
-    DB: "luckshimi"
+  HOST: "localhost",
+  USER: "root",
+  PASSWORD: "luckshimi2!W",
+  PORT: 3306,
+  DB: "luckshimi",
 };
 
 export default dbConfig;
