@@ -1,7 +1,7 @@
-import app from './index';
+import app from "./index";
 
 const PORT = process.env.PORT || 3000;
 
 app.listen(PORT, () => {
-    console.log(`app is running on PORT ${PORT}`)
+  console.log(`app is running on PORT ${PORT}`);
 });
