@@ -1,14 +1,27 @@
-import React from 'react';
-import './about.scss';
+import React from "react";
+import "./about.scss";
 
 function About() {
   return (
     <div className="about">
       <div className="title title_about">Жемчуг</div>
       <div className="about__text">
-        <p className="desctop">Если вы хотите стать владелицей жемчужного украшения, обратите внимание на жемчуг, выбрать и купить который можно не выходя из дома в ювелирном интернет-магазине LUCKSHIMI.</p>
-        <p className="desctop">В ювелирных изделиях собственного производства мы используем только натуральный речной и морской жемчуг в сочетании с благородным серебром и золотом. В большинстве изделий использован речной жемчуг, отличающийся ровной поверхностью, приятными естественными цветами и разнообразием форм.</p>
-        <p>Мы предлагаем вашему вниманию украшения из жемчуга на все случаи жизни. В нашем магазине вы найдете:</p>
+        <p className="desctop">
+          Если вы хотите стать владелицей жемчужного украшения, обратите
+          внимание на жемчуг, выбрать и купить который можно не выходя из дома в
+          ювелирном интернет-магазине LUCKSHIMI.
+        </p>
+        <p className="desctop">
+          В ювелирных изделиях собственного производства мы используем только
+          натуральный речной и морской жемчуг в сочетании с благородным серебром
+          и золотом. В большинстве изделий использован речной жемчуг,
+          отличающийся ровной поверхностью, приятными естественными цветами и
+          разнообразием форм.
+        </p>
+        <p>
+          Мы предлагаем вашему вниманию украшения из жемчуга на все случаи
+          жизни. В нашем магазине вы найдете:
+        </p>
         <ul>
           <li>кольца с жемчугом;</li>
           <li>серьги с жемчугом;</li>
@@ -17,7 +30,10 @@ function About() {
           <li>жемчужные ожерелья;</li>
           <li>комплекты свадебных украшений.</li>
         </ul>
-        <p>Ни одну даму не оставят равнодушной украшения LUCKSHIMI, созданные с учетом последних трендов высокой ювелирной моды.</p>
+        <p>
+          Ни одну даму не оставят равнодушной украшения LUCKSHIMI, созданные с
+          учетом последних трендов высокой ювелирной моды.
+        </p>
       </div>
     </div>
   );
