@@ -1,10 +1,8 @@
-import React from 'react';
-import './copyright.scss';
+import React from "react";
+import "./copyright.scss";
 
 function Copyright() {
-    return (
-        <div className="copyright">LUCKSHIMI, 2012 — 2020</div>
-    )
+  return <div className="copyright">LUCKSHIMI, 2012 — 2020</div>;
 }
 
 export default Copyright;
