@@ -1,3 +1,3 @@
-import Copyright from './Copyright';
+import Copyright from "./Copyright.jsx";
 
 export default Copyright;
