@@ -1,3 +1,3 @@
-import News from './News';
+import News from "./News.jsx";
 
 export default News;
