@@ -1,25 +1,21 @@
-import React from 'react';
-import './phone.scss';
-import Button from '../Button';
+import React from "react";
+import "./phone.scss";
+import Button from "../Button";
 
-//function Phone() {
-class Phone extends React.Component {
+function Phone() {
+  const onOpenCall = () => {
+    this.props.onOpenCall();
+    // console.log(this.props);
+  };
 
-    onOpenCall = () => {
-        this.props.onOpenCall();
-        //console.log(this.props);
-    };
-
-    render() {
-        return (
-            <div className="phone">
-                <span>+7 (499) 391-68-56</span>
-                <Button className='button_call' onClick={this.onOpenCall}>
-                    заказать звонок
-                </Button>
-            </div>
-        )
-    }
+  return (
+    <div className="phone">
+      <span>+7 (499) 391-68-56</span>
+      <Button className="button_call" onClick={onOpenCall}>
+        заказать звонок
+      </Button>
+    </div>
+  );
 }
 
 export default Phone;
