@@ -1,3 +1,3 @@
-import Sales from './Sales';
+import Sales from "./Sales.jsx";
 
 export default Sales;
