@@ -1,12 +1,12 @@
-import React from 'react';
-import './slider.scss';
+import React from "react";
+import "./slider.scss";
 
 function Slider() {
-    return (
-        <div className="slider">
-            <div className="slider__item"></div>
-        </div>
-    )
+  return (
+    <div className="slider">
+      <div className="slider__item"></div>
+    </div>
+  );
 }
 
 export default Slider;
