@@ -1,3 +1,3 @@
-import Textarea from './Textarea';
+import Textarea from "./Textarea.jsx";
 
 export default Textarea;
