@@ -1,23 +1,23 @@
-import React from 'react';
-import './mainPage.scss';
-import Slider from '../../components/Slider';
-import Sales from '../../components/Sales';
-import News from '../../components/News';
-import About from '../../components/About';
-import Feedback from '../../components/Feedback';
-import Map from '../../components/Map';
+import React from "react";
+import "./mainPage.scss";
+import Slider from "../../components/Slider";
+import Sales from "../../components/Sales";
+import News from "../../components/News";
+import About from "../../components/About";
+import Feedback from "../../components/Feedback";
+import Map from "../../components/Map";
 
 function MainPage() {
-    return (
-        <>
-            <Slider />
-            <Sales />
-            <News />
-            <About />
-            <Feedback />
-            <Map />
-        </>
-    )
+  return (
+    <>
+      <Slider />
+      <Sales />
+      <News />
+      <About />
+      <Feedback />
+      <Map />
+    </>
+  );
 }
 
 export default MainPage;
